Rewrite child-script-always-success
