+++ ckb-debugger-tests/Makefile
@@ -2,26 +2,15 @@
 CKB_DEBUGGER ?= ckb-debugger-2023
 
 all: \
-	always-success \
 	cl-always-success \
-	cl-always-success-v2 \
-	cl-always-success-info-cell \
 	cl-child-script \
 	child-script-success \
+	child-script-multisig-success \
+	child-script-multi-inputs \
 	auth-script-exec-success
-
-
-always-success:
-	cargo run --bin always-success |  ${CKB_DEBUGGER} --tx-file=- -s lock
 
 cl-always-success:
 	cargo run --bin cl-always-success |  ${CKB_DEBUGGER} --tx-file=- -s lock
-
-cl-always-success-v2:
-	cargo run --bin cl-always-success-v2 |  ${CKB_DEBUGGER} --tx-file=- -s lock
-
-cl-always-success-info-cell:
-	cargo run --bin cl-always-success-info-cell |  ${CKB_DEBUGGER} --tx-file=- -s lock
 
 cl-child-script:
 	cargo run --bin cl-child-script |  ${CKB_DEBUGGER} --tx-file=- -s lock
