--- ckb-debugger-tests/templates/cl-always-success-v2.json
@@ -1,62 +1,0 @@
-{
-  "mock_info": {
-    "inputs": [
-      {
-        "output": {
-          "capacity": "0x10000000",
-          "lock": {
-            "args": "0x",
-            "code_hash": "0x{{ ref_type combine_lock }}",
-            "hash_type": "type"
-          },
-          "type": null
-        },
-        "data": "0x"
-      }
-    ],
-    "cell_deps": [
-      {
-        "output": {
-          "capacity": "0x10000000",
-          "lock": {
-            "args": "0x",
-            "code_hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
-            "hash_type": "data1"
-          },
-          "type": "{{ def_type combine_lock }}"
-        },
-        "data": "0x{{ data ../../target/riscv64imac-unknown-none-elf/release/ckb-combine-lock-v2 }}"
-      },
-      {
-        "output": {
-          "capacity": "0x10000000",
-          "lock": {
-            "args": "0x",
-            "code_hash": "0x0000000000000000000000000000000000000000000000000000000000000000",
-            "hash_type": "data1"
-          },
-          "type": "{{ def_type always_success }}"
-        },
-        "data": "0x{{ data ../../target/riscv64imac-unknown-none-elf/release/child-script-always-success-v2 }}"
-      }
-    ]
-  },
-  "tx": {
-    "outputs": [
-      {
-        "capacity": "0x0",
-        "lock": {
-          "args": "0x",
-          "code_hash": "0x{{ ref_type always_success }}",
-          "hash_type": "type"
-        }
-      }
-    ],
-    "witnesses": [
-      "0x"
-    ],
-    "outputs_data": [
-      "0x"
-    ]
-  }
-}
