+++ ckb-debugger-tests/templates/cl-always-success.json
@@ -25,7 +25,7 @@
           },
           "type": "{{ def_type combine_lock }}"
         },
-        "data": "0x{{ data ../../target/riscv64imac-unknown-none-elf/release/ckb-combine-lock }}"
+        "data": "0x{{ data ../../target/riscv64imac-unknown-none-elf/release/ckb-combine-lock-v2 }}"
       },
       {
         "output": {
