--- contracts/child-script-always-success-v2/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "child-script-always-success-v2"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-ckb-combine-lock-common = { path = "../../ckb-combine-lock-common" }
-ckb-std = "0.13"
-log = { version = "0.4.17", default-features = false }
