--- contracts/child-script-always-success-v2/src/entry.rs
@@ -1,7 +1,0 @@
-use crate::error::Error;
-use core::result::Result;
-
-pub fn main() -> Result<(), Error> {
-    log::info!("always success!");
-    Ok(())
-}
