--- contracts/child-script-always-success/.cargo/config
@@ -1,2 +1,0 @@
-[build]
-target = "riscv64imac-unknown-none-elf"
