+++ contracts/child-script-always-success/src/main.rs
@@ -7,6 +7,10 @@
 #![no_std]
 #![cfg_attr(not(test), no_main)]
 
+// define modules
+mod entry;
+mod error;
+
 #[cfg(test)]
 extern crate alloc;
 
@@ -17,23 +21,14 @@ ckb_std::entry!(program_entry);
 #[cfg(not(test))]
 default_alloc!();
 
-mod entry;
-mod error;
-
 use ckb_combine_lock_common::logger;
-use log::warn;
 
 /// program entry
-///
-///  Both `argc` and `argv` can be omitted.
-fn program_entry() -> i8 {
+pub fn program_entry() -> i8 {
     drop(logger::init());
     // Call main function and return error code
     match entry::main() {
         Ok(_) => 0,
-        Err(err) => {
-            warn!("script exit with error: {:?}", err);
-            err as i8
-        }
+        Err(err) => err as i8,
     }
 }
